+++ src/app/components/share-show/dc-preview.js
@@ -3,12 +3,14 @@
  * modes.
  *
  * @author Jakub Liput
- * @copyright (C) 2021 ACK CYFRONET AGH
+ * @copyright (C) 2024 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 import { get, computed } from '@ember/object';
+import { bool } from '@ember/object/computed';
 import dublinCoreXmlParser from 'oneprovider-gui/utils/dublin-core-xml-parser';
 import Dc from './-dc';
+import insufficientPrivilegesMessage from 'onedata-gui-common/utils/i18n/insufficient-privileges-message';
 
 export default Dc.extend({
   classNames: ['share-show-dc-preview'],
@@ -28,6 +30,12 @@ export default Dc.extend({
   onChangeEditMode: undefined,
 
   /**
+   * @virtual optional Needed in private views.
+   * @type {Models.Space}
+   */
+  space: undefined,
+
+  /**
    * For format reference see `util:dublin-core-xml-generator#groupedEntries`.
    * @override
    * @type {Array<{ type: String, value: String }>}
@@ -38,6 +46,26 @@ export default Dc.extend({
       'groupedEntries'
     );
   }),
+
+  /**
+   * @type {ComputedProperty<boolean>}
+   */
+  hasManageSharesPrivilege: bool('space.privileges.manageShares'),
+
+  isModifyDisabled: bool('modifyDisabledTip'),
+
+  modifyDisabledTip: computed(
+    'hasManageSharesPrivilege',
+    function modifyDisabledTip() {
+      if (!this.hasManageSharesPrivilege) {
+        return insufficientPrivilegesMessage({
+          i18n: this.i18n,
+          modelName: 'space',
+          privilegeFlag: 'space_manage_shares',
+        });
+      }
+    }
+  ),
 
   actions: {
     startModify() {