+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-12469 Improved UX of managing share name and Open Data title; disabled some operations when user has insufficient privileges for space shares
 * VFS-12582 Added tooltips to file ID column header and label in file details and added share ID
 * VFS-12249 Improved UX for adding/modification xattrs column and fixed headers in others browser
 * VFS-12370 Improved loading indicators by showing loading Oneprovider name